% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_result_parsing.R
\name{barcode_table_to_matrix}
\alias{barcode_table_to_matrix}
\title{Convert an HTO barcode table to a sparse matrix of hto_barcode x cell_barcode}
\usage{
barcode_table_to_matrix(dt, valid_htos)
}
\arguments{
\item{dt}{The data.table of barcode counting results. Must have columns count, cell_barcode, and hto_barcode.}

\item{valid_htos}{A character vector of valid HTO barcode sequences.}
}
\value{
A dgCMatrix with cell barcodes as columns and valid HTO sequences as rows.
}
\description{
Convert an HTO barcode table to a sparse matrix of hto_barcode x cell_barcode
}
\seealso{
fuzzy_filtering
}
