% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_matrix_analysis.R
\name{binarize_hash_matrix}
\alias{binarize_hash_matrix}
\title{Binarize a vector of hash counts}
\usage{
binarize_hash_matrix(mat, cutoff_vals = NULL)
}
\arguments{
\item{mat}{A matrix or sparse matrix of hash tag oligo counts. Rows must be hashes, columns must be cells}

\item{cutoff_vals}{(optional) a named vector of cutoff values to apply. Lenght must be equal to nrow(mat), and names must exist in rownames(mat).}
}
\value{
A list containing two objects:
\itemize{
  \item bmat: a binary sparse matrix (dgCMatrix) with dimensions equal to mat
  \item bsummary: a data.frame with summary values for each hash tag oligo: hash_barcode, cutoff, n_pos, n_neg, frac_pos, frac_neg
}
}
\description{
If cutoff values are provided, they will be used. If not, cutoffs will be determined
using select_hash_cutoff()
}
