% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_hash_key_str}
\alias{check_hash_key_str}
\title{Simple function to check if structure of hashtag key input is correct}
\usage{
check_hash_key_str(in_hto_key)
}
\arguments{
\item{x}{hto_key table with hashtag barcode and id}
}
\value{
no return
}
\description{
Simple function to check if structure of hashtag key input is correct
}
