% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awk_result_parsing.R
\name{fuzzy_filtering}
\alias{fuzzy_filtering}
\title{Fuzzy Filtering for HTO tables}
\usage{
fuzzy_filtering(dt, match_column = "hto_barcode", match_values,
  max_distance = 1)
}
\arguments{
\item{dt}{The data.table to filter. Must have columns count, cell_barcode, and hto_barcode.}

\item{match_column}{A character object matching the column name of the data.table to use for filtering.}

\item{match_values}{A character vector of values to match}

\item{max_distance}{The maximum number of substitutions allowed between values in match_column and any value
in match_values.}
}
\value{
A data.table with the same columns as dt. Non-matching values will be removed, and rows matching the same
value will be merged, with counts summed.
}
\description{
Filters a data.table based on a specific column with imperfect matches to a reference set of values.
This is useful for filtering a table of hash oligo counts where there may be off-by-one errors in the
hash sequences.
}
